Support adding data disks to workers in Azure.
When using local storage, or something like [StorageOS](https://storageos.com),
there might be a need to add one or more data disks to the workers. With this patch
the `azure/container-linux/kubernetes/workers/` module supports a new variable
`data_disks` which allows for one or more volumes to be mounted in the worker
like this:

    module "azure-monitoring-low-priority-pool" {
      source = "git::https://github.com/poseidon/typhoon//azure/container-linux/kubernetes/workers?ref=v1.15.1"
      ...

      data_disks = [{
        disk_size  = 500
        disk_type  = "StandardSSD_LRS"
        mount_path = "/media/volume"
      }]
    }

The list can also be passed via the top level azure module as `worker_data_disks`.