[docs] fix broken link on landing page
The link for "Fedora CoreOS on Google Cloud" was pointing to the wrong place, producing a 404. I clicked it because it said `stable` as opposed to Flatcar Linux's `beta`. I got a good chuckle out of the latter working while stable doesn't.

This PR fixes that. Have a nice day :rocket: