Added support for enabling SSH or not through an option
Currently, `terraform apply` has to be run after the nodes have booted, which means that if one node reboots, human intervention will be required to run `terraform apply` again, or else the node won't be instanced and will stay empty, without the information necessary for it to join the cluster again.
This commit aims to change that through the implementation of an "enable_SSH" option, which defaults to true. Disabling this option will put the files originally distributed through SSH in the butane file, which means they will be included in the ignition.json file and so the machine will go get them just like the other instructions in the ignition file.

short variable name unicity fix