Added an optional extra variable (a map) for machine identification
A MAC address is only composed of 48 bits of which the first 24 bits can be determined as they are the OUI, which really only leaves about 24 random bits.
Furthermore, a MAC address can easily be spoofed, which means using only the MAC for machine identification may become a vulnerability.
This commit concerns the bare-metal/flatcar-linux folder and adds an optional field in the controllers and workers variables as well as a variable in the worker/variables.tf file. It defaults on {}. For example, it can be used to add a field based on the machine's uuid, which is 128 bits long, and thus make bruteforcing much harder (48 bits --> 48 * 128 bits).