os_architecture option added
Until now, typhoon with flatcar-linux on bare metal was only available for machines with an AMD64 (x86_64) architecture. This means there is no currently no support of ARM64 (aarch64) machines.
In this commit, the support for different architectures on flatcar-linux with bare-metal was added. This was done through the addition of an os_architecture variable which replaces the hardcoding of the "amd64" string in some places, such as in the links to the flatcar-linux releases